Modularise fractal config
This makes the fractal config reusable across the various ways we're looking to consume fractal; some freatures added/needed:
-  adds a callback so we can know when fractal has fully strapped and `fractal.components` is ready
- we can ask for a build or server instatation
- pass in where fractal should look for components

This means we'll be able to: `const vfFractalBuild = require('@visual-framework/vf-core/fractal.js).initialize('server',fractalReadyCallback);` in things ranging from Eleventy to child pattern libraries.
